move Target consistency check into separate file
